+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ###head
 
 * improve [intrinsic scaling and move it to a plugin](plugins/intrinsic-dimension)
+* allow use of h descriptor (allow futureproof markup)
 
 ###1.0.0